+++ Web/src/service/firebase/index.js
@@ -1,14 +1,12 @@
-import history from './history'
 import {
     databaseUserRestaurant,
     getUserRestuarant,
     registerUserRestaurant,
     referenceUser,
     referenceUserRestaurant,
-} from './firebase'
+} from './serviceUser'
 
 export {
-    history,
     databaseUserRestaurant,
     getUserRestuarant,
     registerUserRestaurant,
